+++ Vector.cc
@@ -60,6 +60,7 @@
 
 #include "Type.h"
 #include "dods-datatypes.h"
+#include "dods-limits.h"
 #include "escaping.h"
 #include "util.h"
 #include "debug.h"
@@ -177,7 +178,7 @@ bool Vector::m_is_cardinal_type() const
  * @return the size of the buffer created.
  * @exception if the Vector's type is not cardinal type.
  */
-unsigned int Vector::m_create_cardinal_data_buffer_for_type(unsigned int numEltsOfType)
+uint64_t Vector::m_create_cardinal_data_buffer_for_type(uint64_t numEltsOfType)
 {
     // Make sure we HAVE a _var, or we cannot continue.
     if (!d_proto) {
@@ -216,7 +217,7 @@ void Vector::m_delete_cardinal_data_buffer()
  *
  */
 template<class CardType>
-void Vector::m_set_cardinal_values_internal(const CardType* fromArray, int numElts)
+void Vector::m_set_cardinal_values_internal(const CardType* fromArray, uint64_t numElts)
 {
     if (numElts < 0) {
         throw InternalErr(__FILE__, __LINE__, "Logic error: Vector::set_cardinal_values_internal() called with negative numElts!");
@@ -1660,7 +1661,7 @@ static bool types_match(Type t, T *cpp_var)
 /** @brief set the value of a byte array */
 
 template <typename T>
-bool Vector::set_value_worker(T *v, int sz)
+bool Vector::set_value_worker(T *v, uint64_t sz)
 {
     if (!v || !types_match(d_proto->type() == dods_enum_c ? static_cast<D4Enum*>(d_proto)->element_type() : d_proto->type(), v))
         return false;
@@ -1669,43 +1670,43 @@ bool Vector::set_value_worker(T *v, int sz)
     return true;
 }
 
-bool Vector::set_value(dods_byte *val, int sz)
+bool Vector::set_value(dods_byte *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_int8 *val, int sz)
+bool Vector::set_value(dods_int8 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_int16 *val, int sz)
+bool Vector::set_value(dods_int16 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_uint16 *val, int sz)
+bool Vector::set_value(dods_uint16 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_int32 *val, int sz)
+bool Vector::set_value(dods_int32 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_uint32 *val, int sz)
+bool Vector::set_value(dods_uint32 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_int64 *val, int sz)
+bool Vector::set_value(dods_int64 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_uint64 *val, int sz)
+bool Vector::set_value(dods_uint64 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_float32 *val, int sz)
+bool Vector::set_value(dods_float32 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(dods_float64 *val, int sz)
+bool Vector::set_value(dods_float64 *val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
@@ -1717,12 +1718,12 @@ bool Vector::set_value(dods_float64 *val, int sz)
  * @return false if the type of the array is neither Str nor Url
  * or val is null, otherwise returns true.
  */
-bool Vector::set_value(string *val, int sz)
+bool Vector::set_value(string *val, uint64_t sz)
 {
     if ((var()->type() == dods_str_c || var()->type() == dods_url_c) && val) {
         d_str.resize(sz);
         d_capacity = sz;
-        for (int t = 0; t < sz; t++) {
+        for (uint64_t t = 0; t < sz; t++) {
             d_str[t] = val[t];
         }
         set_length(sz);
@@ -1735,60 +1736,60 @@ bool Vector::set_value(string *val, int sz)
 }
 
 template<typename T>
-bool Vector::set_value_worker(vector<T> &v, int sz)
+bool Vector::set_value_worker(vector<T> &v, uint64_t sz)
 {
     return set_value(&v[0], sz);
 }
 
-bool Vector::set_value(vector<dods_byte> &val, int sz)
+bool Vector::set_value(vector<dods_byte> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_int8> &val, int sz)
+bool Vector::set_value(vector<dods_int8> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_int16> &val, int sz)
+bool Vector::set_value(vector<dods_int16> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_uint16> &val, int sz)
+bool Vector::set_value(vector<dods_uint16> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_int32> &val, int sz)
+bool Vector::set_value(vector<dods_int32> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_uint32> &val, int sz)
+bool Vector::set_value(vector<dods_uint32> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_int64> &val, int sz)
+bool Vector::set_value(vector<dods_int64> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_uint64> &val, int sz)
+bool Vector::set_value(vector<dods_uint64> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_float32> &val, int sz)
+bool Vector::set_value(vector<dods_float32> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
-bool Vector::set_value(vector<dods_float64> &val, int sz)
+bool Vector::set_value(vector<dods_float64> &val, uint64_t sz)
 {
     return set_value_worker(val, sz);
 }
 
 
 /** @brief set the value of a string or url array */
-bool Vector::set_value(vector<string> &val, int sz)
+bool Vector::set_value(vector<string> &val, uint64_t sz)
 {
     if (var()->type() == dods_str_c || var()->type() == dods_url_c) {
         d_str.resize(sz);
         d_capacity = sz;
-        for (int t = 0; t < sz; t++) {
+        for (uint64_t t = 0; t < sz; t++) {
             d_str[t] = val[t];
         }
         set_length(sz);
