+++ util.h
@@ -61,7 +61,7 @@ string extract_string_argument(BaseType *arg) ;
 double extract_double_value(BaseType *arg) ;
 double *extract_double_array(Array *a) ;
 void extract_double_array(Array *a, vector<double> &dest) ;
-void set_array_using_double(Array *dest, double *src, int src_len) ;
+void set_array_using_double(Array *dest, double *src, uint64_t src_len) ;
 
 bool is_host_big_endian();
